Use `int` for retrieval and dispersal metrics

+++ api/clients/v2/payloadretrieval/relay_payload_retriever.go
@@ -88,7 +88,7 @@ func (pr *RelayPayloadRetriever) GetPayload(
 		return nil, coretypes.ErrBlobDecodingFailedDerivationError.WithMessage(err.Error())
 	}
 
-	pr.metrics.RecordPayloadSizeBytes(uint(len(payload)))
+	pr.metrics.RecordPayloadSizeBytes(len(payload))
 
 	return payload, nil
 }
